Fix hover behavior and search view
